add user

+++ app/Http/Controllers/TasksController.php
@@ -15,10 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザのタスクの一覧を作成日時の降順で取得）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // Welcomeビューでそれらを表示
+        return view('tasks.index', $data);
     }
 
     /**
@@ -48,6 +59,7 @@ class TasksController extends Controller
         ]);
         
         $task = new Task;
+        $task->user_id = $request->user()->id;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -64,9 +76,14 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::findOrFail($id);
-        return view('tasks.show', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
             'task' => $task,
-        ]);
+            ]);
+        }
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -78,9 +95,14 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
-        return view('tasks.edit', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
-        ]);
+            ]);
+        }
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -92,16 +114,22 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
+        $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            $request->validate([
             'status' => 'required|max:10',
             'content' => 'required|max:255',
-        ]);
+            ]);
         
-        $task = Task::findOrFail($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        return redirect('/');
+            $task = Task::findOrFail($id);
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+            return redirect('/');
+        }
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -113,8 +141,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
-
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/');
+        }
+        else{
+            return redirect('/');
+        }
     }
 }
